flask/requirements.txt
Flask
plotly
dash
cassandra-driver
pandas
geopandas